msvc fixes, simplification and bug-fixes of mp_int, more test coverage (#159)
* fixes broken msvc build
* smaller xml files for testing serialization
* fixing many warnings on msvc
* bug-fixes to mp_int regarding operators and new tests to check their behavior 
* don't expect iterable histogram::at to have size method
* build benchmarks in release mode
* test everything by default when b2 is run on project directory
* more test coverage

+++ test/storage_adaptor_test.cpp
@@ -159,7 +159,7 @@ int main() {
 
   mixed_tests<storage_adaptor<std::vector<int>>,
               storage_adaptor<std::array<double, 100>>>();
-  mixed_tests<unlimited_storage<>, storage_adaptor<std::vector<int>>>();
+  mixed_tests<unlimited_storage<>, storage_adaptor<std::vector<double>>>();
   mixed_tests<storage_adaptor<std::vector<int>>, unlimited_storage<>>();
   mixed_tests<storage_adaptor<std::vector<int>>,
               storage_adaptor<std::map<std::size_t, int>>>();
